+++ airo-camera-toolkit/airo_camera_toolkit/cameras/zed/zed2i.py
@@ -20,18 +20,19 @@ except AssertionError:
 
 import time
 
+import cv2
 import numpy as np
 from airo_camera_toolkit.interfaces import StereoRGBDCamera
 from airo_camera_toolkit.utils.image_converter import ImageConverter
 from airo_typing import (
     CameraIntrinsicsMatrixType,
     CameraResolutionType,
-    ColoredPointCloudType,
     HomogeneousMatrixType,
     NumpyDepthMapType,
     NumpyFloatImageType,
     NumpyIntImageType,
     OpenCVIntImageType,
+    PointCloud,
 )
 
 
@@ -149,7 +150,10 @@ class Zed2i(StereoRGBDCamera):
         self.image_matrix = sl.Mat()
         self.depth_image_matrix = sl.Mat()
         self.depth_matrix = sl.Mat()
-        self.pointcloud_matrix = sl.Mat()
+        self.point_cloud_matrix = sl.Mat()
+
+        self.confidence_matrix = sl.Mat()
+        self.confidence_map = None
 
     @property
     def resolution(self) -> CameraResolutionType:
@@ -188,7 +192,7 @@ class Zed2i(StereoRGBDCamera):
 
     @property
     def depth_enabled(self) -> bool:
-        """Runtime parameter to enable/disable the depth & pointcloud computation. This speeds up the RGB image capture."""
+        """Runtime parameter to enable/disable the depth & point_cloud computation. This speeds up the RGB image capture."""
         return self.runtime_params.enable_depth
 
     @depth_enabled.setter
@@ -218,9 +222,10 @@ class Zed2i(StereoRGBDCamera):
         else:
             view = sl.VIEW.LEFT
         self.camera.retrieve_image(self.image_matrix, view)
-        image: OpenCVIntImageType = self.image_matrix.get_data()
-        image = image[..., :3]  # remove alpha channel
-        image = image[..., ::-1]  # convert from BGR to RGB
+        image_bgra: OpenCVIntImageType = self.image_matrix.get_data()
+        # image = image[..., :3]  # remove alpha channel
+        # image = image[..., ::-1]  # convert from BGR to RGB
+        image = cv2.cvtColor(image_bgra, cv2.COLOR_BGRA2RGB)
         return image
 
     def _retrieve_depth_map(self) -> NumpyDepthMapType:
@@ -238,31 +243,31 @@ class Zed2i(StereoRGBDCamera):
         image = image[..., :3]
         return image
 
-    def get_colored_point_cloud(self) -> ColoredPointCloudType:
+    def _retrieve_colored_point_cloud(self) -> PointCloud:
+        assert self.depth_mode != self.NONE_DEPTH_MODE, "Cannot retrieve depth data if depth mode is NONE"
+        assert self.depth_enabled, "Cannot retrieve depth data if depth is disabled"
+        self.camera.retrieve_measure(self.point_cloud_matrix, sl.MEASURE.XYZ)
+        # shape (width, height, 4) with the 4th dim being x,y,z,(rgba packed into float)
+        # can be nan,nan,nan, nan (no point in the point_cloud on this pixel)
+        # or x,y,z, nan (no color information on this pixel??)
+        # or x,y,z, value (color information on this pixel)
+
+        point_cloud = self.point_cloud_matrix.get_data()
+        points = point_cloud[:, :, :3].reshape(-1, 3)
+        colors = self._retrieve_rgb_image_as_int().reshape(-1, 3)
+
+        return PointCloud(points, colors)
+
+    def _retrieve_confidence_map(self) -> NumpyFloatImageType:
+        self.camera.retrieve_measure(self.confidence_matrix, sl.MEASURE.CONFIDENCE)
+        return self.confidence_matrix.get_data()  # single channel float32 image
+
+    def get_colored_point_cloud(self) -> PointCloud:
         assert self.depth_mode != self.NONE_DEPTH_MODE, "Cannot retrieve depth data if depth mode is NONE"
         assert self.depth_enabled, "Cannot retrieve depth data if depth is disabled"
 
         self._grab_images()
-        self.camera.retrieve_measure(self.pointcloud_matrix, sl.MEASURE.XYZRGBA)
-        # shape (width, height, 4) with the 4th dim being x,y,z,(rgba packed into float)
-        # can be nan,nan,nan, nan (no point in the pointcloud on this pixel)
-        # or x,y,z, nan (no color information on this pixel??)
-        # or x,y,z, value (color information on this pixel)
-
-        # filter out all that have nan in any of the positions of the 3th dim
-        # and reshape to (width*height, 4)
-        point_cloud = self.pointcloud_matrix.get_data()
-        point_cloud = point_cloud[~np.isnan(point_cloud).any(axis=2), :]
-
-        # unpack the colors, drop alpha channel and convert to 0-1 range
-        points = point_cloud[:, :3]
-        colors = point_cloud[:, 3]
-        rgba = np.ravel(colors).view(np.uint8).reshape(-1, 4)
-        rgb = rgba[:, :3]
-        rgb_float = rgb.astype(np.float32) / 255.0  # convert to 0-1 range
-
-        colored_pointcloud = np.concatenate((points, rgb_float), axis=1)
-        return colored_pointcloud
+        return self._retrieve_colored_point_cloud()
 
     @staticmethod
     def list_camera_serial_numbers() -> List[str]:
@@ -297,7 +302,7 @@ if __name__ == "__main__":
     # test rgbd stereo camera
 
     with Zed2i(Zed2i.RESOLUTION_2K, fps=15, depth_mode=Zed2i.PERFORMANCE_DEPTH_MODE) as zed:
-        print(zed.get_colored_point_cloud()[0])  # TODO: test the pointcloud more explicity?
+        print(zed.get_colored_point_cloud().points)  # TODO: test the point_cloud more explicity?
         manual_test_stereo_rgbd_camera(zed)
 
     # profile rgb throughput, should be at 60FPS, i.e. 0.017s